Controllers

+++ OatMilk.Backend.Api/Services/Models/Responses/ModifierResponse.cs
@@ -8,7 +8,7 @@ namespace OatMilk.Backend.Api.Services.Models.Responses
         public Guid Id { get; set; }
         public string Attribute { get; set; }
         
-        public float Magnitude { get; set; }
+        public double Magnitude { get; set; }
         
         public ModifierOperation Operation { get; set; }
     }
